moving towards size_t for str

+++ src/eskilib/eskilib_file.c
@@ -1,7 +1,7 @@
 #include "eskilib_file.h"
 
 //simple fgets implementation that returns the number of characters read
-int_fast32_t eskilib_fgets(char* input_buffer, int size_of_input_buffer, FILE* file_pointer) {
+int_fast32_t eskilib_fgets(char* input_buffer, size_t size_of_input_buffer, FILE* file_pointer) {
 	register int character;
 	register char* buffer = input_buffer;
 	int_fast32_t characters_read = 0;
@@ -19,7 +19,7 @@ int_fast32_t eskilib_fgets(char* input_buffer, int size_of_input_buffer, FILE* f
 	return characters_read;
 }
 
-int_fast32_t eskilib_fgets_delimited(char* input_buffer, int size_of_input_buffer, FILE* file_pointer, char delimiter) {
+int_fast32_t eskilib_fgets_delimited(char* input_buffer, size_t size_of_input_buffer, FILE* file_pointer, char delimiter) {
 	register int character;
 	register char* buffer = input_buffer;
 	int_fast32_t characters_read = 0;
